refactor: feign 모듈 경로 수정
