+++ apps/commerce-api/src/main/java/com/loopers/application/payment/PaymentFacade.java
@@ -1,5 +1,6 @@
 package com.loopers.application.payment;
 
+import com.loopers.domain.notification.NotificationService;
 import com.loopers.domain.order.Order;
 import com.loopers.domain.order.OrderService;
 import com.loopers.domain.payment.*;
