+++ apps/commerce-api/src/main/java/com/loopers/domain/notification/NotificationService.java
@@ -1,4 +1,4 @@
-package com.loopers.application.payment;
+package com.loopers.domain.notification;
 
 import com.loopers.domain.payment.PaymentCommand;
 import lombok.extern.slf4j.Slf4j;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @Service
 public class NotificationService {
-    void sendPaymentSyncFailureAlert(PaymentCommand.CallbackRequest command) {
+    public void sendPaymentSyncFailureAlert(PaymentCommand.CallbackRequest command) {
         log.info("Sending payment sync failure alert for payment command {}", command);
     }
 }
