+++ modules/feign/src/main/resources/feign.yml
